Extract prefs models
