+++ features/bible/src/main/kotlin/app/ss/bible/BibleVersesActivity.kt
@@ -27,8 +27,6 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import app.ss.bible.databinding.SsBibleVersesActivityBinding
-import app.ss.models.SSReadingDisplayOptions
-import app.ss.models.colorTheme
 import com.cryart.sabbathschool.core.extensions.context.isDarkTheme
 import com.cryart.sabbathschool.core.extensions.coroutines.flow.collectIn
 import com.cryart.sabbathschool.core.extensions.view.viewBinding
@@ -36,6 +34,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.map
 import ss.misc.SSConstants
 import ss.misc.SSEvent.track
+import ss.prefs.model.SSReadingDisplayOptions
+import ss.prefs.model.colorTheme
 
 @AndroidEntryPoint
 class BibleVersesActivity : AppCompatActivity(), ToolbarComponent.Callbacks {
