+++ features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/readings/options/SSReadingDisplayOptionsViewModel.kt
@@ -22,9 +22,9 @@
 package com.cryart.sabbathschool.lessons.ui.readings.options
 
 import androidx.core.view.children
-import app.ss.models.SSReadingDisplayOptions
 import com.cryart.sabbathschool.lessons.databinding.SsReadingDisplayOptionsBinding
 import ss.prefs.api.SSPrefs
+import ss.prefs.model.SSReadingDisplayOptions
 
 class SSReadingDisplayOptionsViewModel(
     private val binding: SsReadingDisplayOptionsBinding,
